Rewrite python generation
The idea of having fixed python configuration that could be
customized via the usual template mechanism is too rigid and
quickly become obsolete with respect to the ongoing
development in release in terms of conditions, GT,
customizations, geometries, etc. For this reason a new
mechanisms has been introduced in order to generate and
customize the python configuration files. A shell script has
been added in order to generate python fragments using
directly cmsDriver.py commands, that should always produce
up-to-date results. The template variables have been
isolated and applied on top of those fragments. The commands
should keep on working as in the past, except that the user
now has to generate the fragments everytime.
One minor point is that the directory name must not contain
the '-' character, otherwise we are not able to import the
generated fragments.